Add support for custom output template for each rootdir
to use: add a file named "template.txt" into the rootdir containing only
the part of the template which comes after the directory. No protection
against escaping from the target directory so this is probably unsafe
AF. Use with caution.

Also rename the second rootdir in the example config file